TERR-425 - removed watch mode from jest react test invocation from package.json - updated playwright.config to run split off of edge coloration tests in seperate module for speed up of parallel execution - split tests into main plugin.spec module and edgeColoration.spec module - renamed parent folder for untypedFn module to util - added fixtures module to avoid repeeated invocations of fixture instatiation code - moved common url related fn for tests into util/urlHelpers module - added check on thresholds in MapPanel to allow for non-array values like null or undefined to bypass forEach invocation - removed unused suffix input param prop in CustomTextArea component - added @ts-ignore flags to minimize TS type error noise from CustomTextArea components in tests due to bad React 17 FC and VFC signatures WRT to children.
