+++ src/components/CustomTextArea.tsx
@@ -1,16 +1,12 @@
-import React, { ReactNode, useCallback, useEffect, useRef } from 'react';
+import React, { useCallback, useEffect, useRef } from 'react';
 
-import { StandardEditorProps, StringFieldConfigSettings } from '@grafana/data';
+import { StringFieldConfigSettings } from '@grafana/data';
 import { TextArea } from '@grafana/ui';
 import { monospacedFontSize } from '../options';
 
 export interface CustomTextAreaSettings extends StringFieldConfigSettings {
   isMonospaced: boolean;
   fontSize: string;
-}
-
-interface Props extends StandardEditorProps<string, CustomTextAreaSettings> {
-  suffix?: ReactNode;
 }
 
 function unescape(str) {
@@ -23,16 +19,18 @@ function unescape(str) {
 
 /**
  * This component renders a customized TextArea that accepts an item defined by Grafana's StandardEditorsRegistryItem
- * interface.
+ * interface. The CustomTextArea differs from Grafana's TextArea in that changes containing HTML entities (for instance,
+ * &amp; &lt; &gt;, etc) are unescaped and thus stored internally unchanged. This is intended for input that is not meant
+ * to be rendered back to the page in HTML.
+ *
+ * Supports the setting 'isMonospaced' to render the text field using a monospaced font, as opposed to a proportional font.
  *
  * @prop {string} value                                           The current value of the standard registry item
  * @prop {(value?: any): void} onChange                           A callback invoked when the value in the text area changes.
  * @prop {StandardEditorsRegistryItem<TValue, TSettings>} item    The standard registry item to edit
- * @prop {TBD} suffix                                             Optional.
- * @param param0
  * @returns
  */
-export const CustomTextArea: React.FC<Props> = ({ value, onChange, item, suffix }) => {
+export const CustomTextArea = ({ value, onChange, item }) => {
   let textareaRef = useRef<HTMLTextAreaElement>(null);
 
   const onValueChange = useCallback(
